+++ mvc/src/main/java/nextstep/mvc/controller/tobe/ControllerScanner.java
@@ -1,0 +1,47 @@
+package nextstep.mvc.controller.tobe;
+
+import java.lang.reflect.Constructor;
+import java.lang.reflect.InvocationTargetException;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Set;
+import nextstep.web.annotation.Controller;
+import org.reflections.Reflections;
+
+public class ControllerScanner {
+
+    private final Reflections reflections;
+
+    private ControllerScanner(final Reflections reflections) {
+        this.reflections = reflections;
+    }
+
+    public static ControllerScanner from(final Object[] basePackages) {
+        return new ControllerScanner(new Reflections(basePackages));
+    }
+
+    public Map<Class<?>, Object> getControllers() {
+        final Set<Class<?>> controllerClasses = reflections.getTypesAnnotatedWith(Controller.class);
+        return instantiateControllers(controllerClasses);
+    }
+
+    private Map<Class<?>, Object> instantiateControllers(final Set<Class<?>> controllerClasses) {
+        final Map<Class<?>, Object> controllers = new HashMap<>();
+        for (Class<?> controllerClass : controllerClasses) {
+            controllers.put(controllerClass, createControllerInstance(controllerClass));
+        }
+
+        return controllers;
+    }
+
+    private Object createControllerInstance(final Class<?> controllerClass) {
+        try {
+            final Constructor<?> declaredConstructor = controllerClass.getDeclaredConstructor();
+            declaredConstructor.setAccessible(true);
+            return declaredConstructor.newInstance();
+        } catch (InvocationTargetException | InstantiationException | IllegalAccessException |
+                 NoSuchMethodException exception) {
+            throw new IllegalStateException("Fail to create handler instance");
+        }
+    }
+}
